LayoutTests/platform/gtk/http/tests/dom/document-attributes-null-handling-expected.txt

Test null handling of several Document attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* cookie attribute
PASS document.cookie is ""
document.cookie = 'key=value'
PASS document.cookie is "key=value"
document.cookie = null
PASS document.cookie.indexOf('key=value') != -1 is true
FAIL document.cookie.indexOf('null=') != -1 should be true. Was false.

* title attribute
PASS document.title is ""
document.title = null
PASS document.title is "null"

* domain attribute
PASS document.domain is "127.0.0.1"
PASS document.domain = null threw exception SecurityError (DOM Exception 18): The operation is insecure..
PASS document.domain is "127.0.0.1"
PASS successfullyParsed is true

TEST COMPLETE
